requirements.txt
# DiDi POI 热点系统依赖
# 用法: pip3 install -r requirements.txt

# Web 框架
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2

# 网络请求
requests==2.31.0
urllib3==2.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.6

# 数据处理
pandas==2.1.3
numpy==2.0.0
openpyxl==3.1.2

# 地图和可视化
folium==0.20.0
branca==0.8.2
Pillow==10.1.0

# 图像处理
opencv-python==4.8.1.78

# API 和 LLM
anthropic==0.25.0

# 环境变量
python-dotenv==1.0.0

# 日期时间处理
pytz==2023.3.post1

# 任务调度
APScheduler==3.10.4

# 加密和安全
cryptography==41.0.7

# 其他实用库
typing-extensions==4.8.0
